Add codecov badge to the project
Codecov is a badge which display how much of the code is covered by tests.

To enable it to your repository, follow the instructions on [codecov page](https://codecov.io/)

The travisCI should then update the status after each individual commit pushed to github, or in each PR sent to it.